requirements.txt
PyQt5
PyOpenGL
numpy